GH-127: Extracted out ConfirmationDialog as a reusable component
