+++ ui/src/components/custom/ConfirmationDialog.tsx
@@ -1,0 +1,52 @@
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogContentText,
+  DialogTitle,
+} from '@mui/material';
+
+interface ConfirmationDialogProps {
+  open: boolean;
+  onClose: () => void;
+  onConfirm: () => void;
+  title: string;
+  message: string;
+  confirmText?: string;
+  cancelText?: string;
+  confirmColor?: 'primary' | 'error' | 'warning' | 'success';
+}
+
+const ConfirmationDialog = ({
+  open,
+  onClose,
+  onConfirm,
+  title,
+  message,
+  confirmText = 'Confirm',
+  cancelText = 'Cancel',
+  confirmColor = 'primary',
+}: ConfirmationDialogProps) => (
+  <Dialog
+    open={open}
+    onClose={onClose}
+    aria-labelledby="confirmation-dialog-title"
+    aria-describedby="confirmation-dialog-description"
+  >
+    <DialogTitle id="confirmation-dialog-title">{title}</DialogTitle>
+    <DialogContent>
+      <DialogContentText id="confirmation-dialog-description">{message}</DialogContentText>
+    </DialogContent>
+    <DialogActions>
+      <Button onClick={onClose} color="primary">
+        {cancelText}
+      </Button>
+      <Button onClick={onConfirm} color={confirmColor} autoFocus>
+        {confirmText}
+      </Button>
+    </DialogActions>
+  </Dialog>
+);
+
+export default ConfirmationDialog;
